Extracts Notice logic into the PluginsNoticeSection component
